feat: dto 분리, 유효성 검사 추가
