+++ backend/src/node/dtos/createNode.dto.ts
@@ -1,0 +1,12 @@
+import { IsString, IsNumber } from 'class-validator';
+
+export class CreateNodeDto {
+  @IsString()
+  title: string;
+
+  @IsNumber()
+  x: number;
+
+  @IsNumber()
+  y: number;
+}
