+++ backend/src/node/dtos/updateNode.dto.ts
@@ -1,0 +1,12 @@
+import { IsString, IsNumber } from 'class-validator';
+
+export class UpdateNodeDto {
+  @IsString()
+  title: string;
+
+  @IsNumber()
+  x: number;
+
+  @IsNumber()
+  y: number;
+}
