--- backend/src/node/node.dto.ts
@@ -1,11 +1,0 @@
-export class CreateNodeDto {
-  title: string;
-  x: number;
-  y: number;
-}
-
-export class UpdateNodeDto {
-  title: string;
-  x: number;
-  y: number;
-}
