+++ backend/src/page/dtos/updatePage.dto.ts
@@ -1,0 +1,9 @@
+import { IsString, IsJSON } from 'class-validator';
+
+export class UpdatePageDto {
+  @IsString()
+  title: string;
+
+  @IsJSON()
+  content: JSON;
+}
