--- backend/src/page/page.dto.ts
@@ -1,11 +1,0 @@
-export class CreatePageDto {
-  title: string;
-  content: JSON;
-  x: number;
-  y: number;
-}
-
-export class UpdatePageDto {
-  title: string;
-  content: JSON;
-}
