+++ backend/src/page/page.service.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { NodeRepository } from '../node/node.repository';
 import { PageRepository } from './page.repository';
 import { Page } from './page.entity';
-import { CreatePageDto, UpdatePageDto } from './page.dto';
+import { CreatePageDto } from './dtos/createPage.dto';
+import { UpdatePageDto } from './dtos/updatePage.dto';
 import { PageNotFoundException } from '../exception/page.exception';
 
 @Injectable()
